nyota-enterprise/core/requirements.txt
fastapi
uvicorn
nats-py
pydantic
python-dotenv
websockets
asyncpg
